first try

+++ modules/domain/src/main/kotlin/dev/h2r/domain/entity/Client.kt
@@ -1,10 +1,9 @@
 package dev.h2r.domain.entity
 
-import java.math.BigInteger
 
 data class Client(
     val id: Int,
-    val balance: BigInteger,
-    val limit: BigInteger,
+    val balance: Int,
+    val limit: Int,
     val last10Transactions: List<Transaction>
 )
