+++ modules/domain/src/main/kotlin/dev/h2r/domain/entity/TransactionResult.kt
@@ -1,8 +1,6 @@
 package dev.h2r.domain.entity
 
-import java.math.BigInteger
-
 data class TransactionResult(
-    val limit: BigInteger,
-    val balance: BigInteger
+    val limit: Int,
+    val balance: Int
 )
