+++ modules/domain/src/main/kotlin/dev/h2r/domain/port/ClientDatabasePort.kt
@@ -2,10 +2,9 @@ package dev.h2r.domain.port
 
 import dev.h2r.domain.entity.Client
 import dev.h2r.domain.entity.Transaction
-import java.math.BigInteger
 
 interface ClientDatabasePort {
-    fun updateBalance(client: Client, transaction: Transaction, newBalance: BigInteger): Client?
+    fun updateBalance(client: Client, transaction: Transaction, newBalance: Int): Client?
     fun find(idClient: Int): Client?
 
 }
